refactor(board): reorganize by sub-domain instead of technical role
Restructure src/features/board/ from technical-role folders
(components/, hooks/, db/, store/, mappers/) to sub-domain folders
that communicate what each group does, not what file type it contains:

- grid/       — visual grid layout, tiles, pictograms, keyboard nav
- message/    — sentence strip: composition, playback, UI controls
- navigation/ — board-to-board routing and nav buttons
- suggestions/— AI-powered phrase suggestions and tone selection
- storage/    — IndexedDB persistence and board-sets state
- import/     — OBZ/OBF file parsing, format mapping, import hooks

Cross-cutting orchestrators remain at the feature root:
BoardView, useBoard, useLoadBoard, useBoardTranslation,
useButtonActivation.

Pure structural change — no functional modifications.
All external consumers import from the barrel (index.ts),
so zero changes outside the feature boundary.
