+++ src/features/board/grid/Pictogram.test.tsx
