+++ src/features/board/grid/Pictogram.tsx
