+++ src/features/board/grid/Tile.tsx
@@ -1,7 +1,7 @@
 import Button from "@mui/material/Button";
 import { darken } from "@mui/material/styles";
 import { getReadableTextColor } from "@shared/utils/colors";
-import { Pictogram } from "../Pictogram/Pictogram";
+import { Pictogram } from "./Pictogram";
 
 export interface TileProps {
   label?: string;
