+++ src/features/board/import/board-import.ts
@@ -1,14 +1,14 @@
 import { stripHtmlTags } from "@shared/utils/html";
 import { lookup } from "mrmime";
 import { loadOBF, loadOBZ, type OBFBoard } from "open-board-format";
-import { resolveLoadBoardPaths } from "../mappers/obf-mapper";
-import type { BoardsDB, UpsertBoardSetInput } from "./boards-db";
+import { resolveLoadBoardPaths } from "./obf-mapper";
+import type { BoardsDB, UpsertBoardSetInput } from "../storage/boards-db";
 import {
   putAssets,
   putBoards,
   upsertBoardSet,
   withBoardsDB,
-} from "./boards-db";
+} from "../storage/boards-db";
 
 export interface ImportResult {
   setId: string;
