+++ src/features/board/index.ts
@@ -1,11 +1,11 @@
-export { BoardView } from "./components/BoardView/BoardView";
-export type { BoardSetRecord } from "./db/boards-db";
-export { useBoard } from "./hooks/useBoard";
-export { useBoardSets } from "./hooks/useBoardSets";
-export { useImportBoardFiles } from "./hooks/useImportBoardFiles";
+export { BoardView } from "./BoardView";
+export type { BoardSetRecord } from "./storage/boards-db";
+export { useBoard } from "./useBoard";
+export { useBoardSets } from "./storage/useBoardSets";
+export { useImportBoardFiles } from "./import/useImportBoardFiles";
 export {
   fetchBoardSets,
   importBoardFromUrl,
   removeBoardSet,
-} from "./store/board-sets-store";
+} from "./storage/board-sets-store";
 export type { BoardRouteParams } from "./types";
