+++ src/features/board/navigation/useBoardNavigation.test.tsx
@@ -2,10 +2,10 @@ import type { ReactNode } from "react";
 import { MemoryRouter, Route, Routes } from "react-router";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { renderHook } from "vitest-browser-react";
-import * as boardSetsStore from "../store/board-sets-store";
+import * as boardSetsStore from "../storage/board-sets-store";
 import { useBoardNavigation } from "./useBoardNavigation";
 
-vi.mock("../store/board-sets-store", () => ({
+vi.mock("../storage/board-sets-store", () => ({
   subscribeBoardSets: vi.fn(),
   getBoardSetsSnapshot: vi.fn(),
 }));
