+++ src/features/board/storage/useBoardSets.ts
@@ -1,9 +1,6 @@
 import { useSyncExternalStore } from "react";
-import type { BoardSetRecord } from "../db/boards-db";
-import {
-  getBoardSetsSnapshot,
-  subscribeBoardSets,
-} from "../store/board-sets-store";
+import type { BoardSetRecord } from "./boards-db";
+import { getBoardSetsSnapshot, subscribeBoardSets } from "./board-sets-store";
 
 export interface UseBoardSetsReturn {
   boardSets: BoardSetRecord[];
