+++ src/features/board/suggestions/SuggestionBar.tsx
@@ -1,7 +1,7 @@
 import Box from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import Stack from "@mui/material/Stack";
-import type { SuggestionTone } from "../../types";
+import type { SuggestionTone } from "../types";
 import { ToneSelector } from "./components/ToneSelector";
 
 export interface SuggestionBarProps {
